TPCH Expts (#25)
* Making TPCH queries run...

* Setup experimiments

* Running TPCH expts...

* Update Dependencies.cmake

Undo inadvertent CMake LLVM Changes.

* Update constant_translator.cpp

Removed duplicate include.

* Update vm.cpp

Fixed comment for consistency.

* Update benchmark_main.cpp

Disable TPCH data loading in benchmark main. Should only be in TPCH benchmark.

* Update benchmark_main.cpp

Cleaned up includes.

+++ src/include/sql/codegen/codegen.h
@@ -460,7 +460,7 @@ class CodeGen {
    * @param num The number to convert.
    * @return The SQL real.
    */
-  [[nodiscard]] ast::Expr *FloatToSql(int64_t num) const;
+  [[nodiscard]] ast::Expr *FloatToSql(double num) const;
 
   /**
    * Call @dateToSql(). Convert a date into a SQL date.
